chore: Organize components into atomic pattern
