+++ babel.config.js
@@ -8,6 +8,9 @@ module.exports = {
         extensions: ['.ios.js', '.android.js', '.js', '.ts', '.tsx', '.json'],
         alias: {
           '@components': './src/components',
+          '@atoms': './src/components/atoms',
+          '@molecules': './src/components/molecules',
+          '@organisms': './src/components/organisms',
           '@store': './src/store',
           '@app-types': './src/types',
           '@routes': './src/routes',
