+++ src/components/organisms/BottomTabs/styles.ts
