+++ tsconfig.json
@@ -4,6 +4,9 @@
     "baseUrl": ".",
     "paths": {
       "@components/*": ["src/components/*"],
+      "@atoms/*": ["src/components/atoms/*"],
+      "@molecules/*": ["src/components/molecules/*"],
+      "@organisms/*": ["src/components/organisms/*"],
       "@store/*": ["src/store/*"],
       "@app-types/*": ["src/types/*"],
       "@routes/*": ["src/routes/*"],
